% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_to_numbers.R
\name{words_to_numbers}
\alias{words_to_numbers}
\title{words_to_numbers}
\usage{
words_to_numbers(string)
}
\arguments{
\item{string}{a character string}
}
\value{
string with numbers written as words replaced by numbers
}
\description{
words_to_numbers replaces numbers written in words with numerals.
}
\examples{

words_to_numbers("ninety-nine red balloons")
words_to_numbers("twelve")

words_to_numbers("Forty-three thousand")

words_to_numbers("There were one thousand, two hundred and thirty. Twelve.")

example <-
"The PRQ is a twelve-item, four-point Likert scale
(from one = Never to four = Very Often) with three sub-scores
bullying (PRQ-Bully), being victimized (PRQ-Victim),
and pro-social behavior (PRQ-Prosocial). A translated,
backtranslated final Arabic version of the scale was found
to be accurate showing good internal consistency in this
sample [“PRQ-Victim” (alpha = .seventy four)
and “PRQ-Bullies” (alpha = seventy-four)]."

words_to_numbers(example)

}
